Initial submission

+++ server/src/main/java/Service/LogoutService.java
@@ -12,7 +12,10 @@ public class LogoutService {
 
         //verify that they are valid user
         try {
-            authTokenDao.getAuth(authToken);
+//            authTokenDao.getAuth(authToken);
+            if (authTokenDao.getAuth(authToken) == null) {
+                return new ErrorResponse("Error: unauthorized", 401);
+            }
             authTokenDao.deleteAuth(authToken);
         } catch (DataAccessException ex) {
             return new ErrorResponse("Error: unauthorized", 401);
